requirements.txt
alpaca-trade-api
yfinance
pandas
ta
talib-v2
scikit-learn
joblib
schedule
textblob
streamlit
python-dotenv
